Eczane gir fonksiyonunda float yerine double değişikliği yapıldı.hastaGecmisGir fonksiyonu silindi.

+++ app/src/main/java/com/example/cihan/eczanebul/Database.java
@@ -189,7 +189,7 @@ public class Database extends SQLiteOpenHelper {
         db.execSQL(String.format(insert,id,isim,mg,kul,tarih,null));
     }
 
-    public void eczaneGir(int id,String sifre,String isim,String adres,String telefon,float koordx,float koordy){
+    public void eczaneGir(int id,String sifre,String isim,String adres,String telefon,double koordx,double koordy){
         SQLiteDatabase db = getWritableDatabase();
         Cursor cursor = db.query("Eczane",null,"ID="+id,null,null,null,null);
         cursor.moveToFirst();
